fix: moving collector package under the main namespace or otherwise spring won't like it
