--- src/main/java/org/rcsb/collectors/alignments/AlignmentBuilder.java
@@ -1,19 +1,0 @@
-/*
- *  Copyright (c) 2024 - now, RCSB PDB and contributors, licensed under MIT, See LICENSE file for more info.
- */
-
-package org.rcsb.collectors.alignments;
-
-import org.bson.Document;
-import org.rcsb.graphqlschema.reference.SequenceReference;
-import reactor.core.publisher.Flux;
-
-/**
- * @author : joan
- * @mailto : joan.segura@rcsb.org
- * @created : 2/5/24, Monday
- **/
-
-public interface AlignmentBuilder {
-    Flux<Document> apply(String queryId, SequenceReference from, SequenceReference to);
-}
